+++ pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   flutter_riverpod: ^2.3.2
   import_sorter: ^4.6.0
+  google_fonts: ^4.0.3
 
 dev_dependencies:
   flutter_test:
